refactor: move the definition of link TRB to `ring` module
This TRB is also used by transfer TRB.
