+++ kernel/src/device/pci/xhci/structures/ring/command/trb.rs
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use super::super::CycleBit;
+use super::super::{CycleBit, Link};
 use crate::add_trb;
 use bit_field::BitField;
-use core::convert::{TryFrom, TryInto};
+use core::convert::TryInto;
 use os_units::Bytes;
 use x86_64::PhysAddr;
 
@@ -45,25 +45,6 @@ impl From<Trb> for [u32; 4] {
             Trb::EnableSlot(e) => e.0,
             Trb::AddressDevice(a) => a.0,
         }
-    }
-}
-
-add_trb!(Link);
-impl Link {
-    const ID: u8 = 6;
-    fn new(addr_to_ring: PhysAddr) -> Self {
-        assert!(addr_to_ring.is_aligned(u64::try_from(Trb::SIZE.as_usize()).unwrap()));
-        let mut trb = Self([0; 4]);
-        trb.set_trb_type(Self::ID);
-        trb.set_addr(addr_to_ring.as_u64());
-        trb
-    }
-
-    fn set_addr(&mut self, a: u64) {
-        let l = a & 0xffff_ffff;
-        let u = a >> 32;
-        self.0[0] = l.try_into().unwrap();
-        self.0[1] = u.try_into().unwrap();
     }
 }
